update to separate current liabilities view
